1.  Refactored CIA implementation in the MAME Amiga code into a separate module
2.  [CBM] Changed code to share 6526 CIA implementation with the one used by
    the Amiga in MAME.

3.  [AMIGA] Misc cleanups

4.  Removed mess/machine/cia6526.c implementation

+++ mess/machine/amigafdc.h
@@ -1,7 +1,7 @@
 #ifndef AMIGAFDC_H
 #define AMIGAFDC_H
 
-void amiga_fdc_control_w( int data );
+void amiga_fdc_control_w( UINT8 data );
 int amiga_fdc_status_r( void );
 unsigned short amiga_fdc_get_byte( void );
 void amiga_fdc_setup_dma( void );
